AI-learning/handson/requirements.txt
openai>=1.30.0
pyyaml>=6.0
# Module 4 のみ必要
mcp>=1.0
